Move REPL diff tests out of `tests.scala`
