+++ compiler/test/dotty/tools/dotc/CompilerTest.scala
@@ -237,23 +237,6 @@ abstract class CompilerTest {
     }
   }
 
-  def replFile(prefix: String, fileName: String): Unit = {
-    val path = s"$prefix$fileName"
-    val f = new PlainFile(path)
-    val repl = new TestREPL(new String(f.toCharArray))
-    repl.process(Array[String]())
-    repl.check()
-  }
-
-  def replFiles(path: String): Unit = {
-    val dir = Directory(path)
-    val fileNames = dir.files.toArray.map(_.jfile.getName).filter(_ endsWith ".check")
-    for (name <- fileNames) {
-      log(s"testing $path$name")
-      replFile(path, name)
-    }
-  }
-
   // ========== HELPERS =============
 
   private def expectedErrors(filePaths: List[String]): List[ErrorsInFile] = if (filePaths.exists(isNegTest(_))) filePaths.map(getErrors(_)) else Nil
